% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALCULATE_DF.R
\name{CALCULATE_DF}
\alias{CALCULATE_DF}
\title{eQC/Test Sample DF Calculator}
\usage{
CALCULATE_DF(site, username, password, analyte, species, matrix, OS_file)
}
\arguments{
\item{site}{The PFS version need to work on: "Test" or "Production"}

\item{username}{The username to log in PFS}

\item{password}{The password to log in PFS}

\item{analyte}{Target analyte name}

\item{species}{Species of the experiment}

\item{matrix}{Matrix of the experiment}

\item{OS_file}{Experiment OS result}
}
\value{
Return two data frame including calculated dilution factor
}
\description{
Calculate normalized dilution factor for different batch
}
\examples{
CALCULATE_DF(site, username, password, analyte, species, matrix, OS_file)
}
