% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHECK_SAMP_ConDup.R
\name{CHECK_SAMP_ConDup}
\alias{CHECK_SAMP_ConDup}
\title{DTA Consistency/Duplication Check}
\usage{
CHECK_SAMP_ConDup(data, dup_cols, std_cols, check_cols)
}
\arguments{
\item{data}{The data frame after data transfer}

\item{dup_cols}{The columns to check duplication}

\item{std_cols}{The columns as conditions}

\item{check_cols}{The columns to check consistency}
}
\value{
Return data frames with inconsistent/duplicated data
}
\description{
Check whether DTA file is consistent/duplication with the selected columns
}
\examples{
CHECK_SAMP_ConDup(data, dup_cols = c("Subject ID", "Visit Name"), std_cols = c("Subject ID", "Visit Name", "Sample Type"), check_cols = c("Nextcea Received Date", "Gender"));
}
