% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHECK_SAMP_Consis.R
\name{CHECK_SAMP_Consis}
\alias{CHECK_SAMP_Consis}
\title{DTA Consistency Check}
\usage{
CHECK_SAMP_Consis(data, std_cols, check_cols)
}
\arguments{
\item{data}{The data frame after data transfer}

\item{std_cols}{The columns as conditions}

\item{check_cols}{The columns to check consistency}
}
\value{
Return data frames with expected inconsistent data
}
\description{
Check whether DTA file is consistent with the selected columns
}
\examples{
CHECK_SAMP_Consis(data, std_cols = c("Subject ID", "Visit Name", "Sample Type"), check_cols = c("Nextcea Received Date", "Gender"));
}
