% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHECK_SAMP_Error.R
\name{CHECK_SAMP_Error}
\alias{CHECK_SAMP_Error}
\title{DTA Error Check}
\usage{
CHECK_SAMP_Error(format, data)
}
\arguments{
\item{format}{The format of DTA data}

\item{data}{The data frame after data transfer}
}
\value{
Return data frames with expected analytes and BQL values
}
\description{
Check DTA file expected analytes and BQL values
}
\examples{
CHECK_SAMP_Error(data);
}
