% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{count.llk}
\alias{count.llk}
\title{Calculate the summation of logP(Y|theta1)}
\usage{
count.llk(nodesLabel, theta_1, Y, N, mu)
}
\arguments{
\item{nodesLabel}{The vector of latent state of all genes.}

\item{theta_1}{The value of theta_1.}

\item{Y}{The vector of DNM counts for all genes.}

\item{N}{The parameter in the prior for network.}

\item{mu}{The vector of mutability for all genes.}
}
\value{
The summation of logP(Y|theta1)
}
\description{
Calculate the summation of logP(Y|theta1)
}
