% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CREATE_ENTITY_AttrValue.R
\name{CREATE_ENETITY_AttrValue}
\alias{CREATE_ENETITY_AttrValue}
\title{Entity Record Create}
\usage{
CREATE_ENETITY_AttrValue(site, username, password, entity, body)
}
\arguments{
\item{site}{The PFS version need to work on: "Test" or "Production"}

\item{username}{The username to log in PFS}

\item{password}{The password to log in PFS}

\item{entity}{The Entity name}

\item{body}{Values in a list to fill the Entity attribute}
}
\value{
Return POST result message
}
\description{
Create Entity Record with Attributes' Value
}
\examples{
CREATE_ENETITY_AttrValue("Test", "user", "password", Entity, body);
}
