% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_LCMS_SampleBarcodes.R
\name{GET_LCMS_SampleBarcodes}
\alias{GET_LCMS_SampleBarcodes}
\title{Sample Barcode Query}
\usage{
GET_LCMS_SampleBarcodes(site, ept_barcode, username, password)
}
\arguments{
\item{site}{The PFS version need to work on: "Test" or "Production"}

\item{ept_barcode}{The Experiment Barcode of samples}

\item{username}{The username to log in PFS}

\item{password}{The password to log in PFS}
}
\value{
Return one data frame including samples information: barcode, name, and active status
}
\description{
Get LCMS Sample Barcode from LIMS
}
\examples{
GET_LCMS_SampleBarcodes("Test", "LCMS1", "user", "password");
}
