% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_PROJ_Range.R
\name{GET_PROJ_Range}
\alias{GET_PROJ_Range}
\title{LLOQ Info Query}
\usage{
GET_PROJ_Range(site = "Test", username, password)
}
\arguments{
\item{site}{The PFS version need to work on: "Test" or "Production", "Test" as default}

\item{username}{The username to log in PFS}

\item{password}{The password to log in PFS}
}
\value{
Return one data frame including species, matrix, LLOQ, ULOQ and other info
}
\description{
Get Project LLOQ Info from PFS
}
\examples{
GET_PROJ_Range("Test", "user", "password");
}
