% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_PROJ_SampleLot.R
\name{GET_PROJ_SampleLot}
\alias{GET_PROJ_SampleLot}
\title{Sample Lot Query}
\usage{
GET_PROJ_SampleLot(site, project, username, password)
}
\arguments{
\item{site}{The PFS version need to work on: "Test" or "Production"}

\item{project}{The Project Barcode to upload the analyte result}

\item{username}{The username to log in PFS}

\item{password}{The password to log in PFS}
}
\value{
Return one data frame including all the sample information in the sample lot
}
\description{
Get Sample Lot Information from PFS
}
\examples{
GET_LCMS_SampleBarcodes("Test", "NP1", "user", "password");
}
