% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_SAMP_ExpInfo.R
\name{GET_SAMP_ExpInfo}
\alias{GET_SAMP_ExpInfo}
\title{Sample Experiment Info Query}
\usage{
GET_SAMP_ExpInfo(site, project, username, password)
}
\arguments{
\item{site}{The PFS version need to work on: "Test" or "Production"}

\item{project}{The Project Barcode to upload the analyte result}

\item{username}{The username to log in PFS}

\item{password}{The password to log in PFS}
}
\value{
Return one data frame including sample index, test date, and LLOQ/ULOQ
}
\description{
Get Sample Experiment Info from PFS
}
\examples{
GET_SAMP_ExpmtInfo("Test", "NP1", "user", "password");
}
