% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_SAMP_TestDate.R
\name{GET_SAMP_TestDate}
\alias{GET_SAMP_TestDate}
\title{Sample Test Date Query}
\usage{
GET_SAMP_TestDate(site, project, username, password)
}
\arguments{
\item{site}{The PFS version need to work on: "Test" or "Production"}

\item{project}{The Project Barcode to upload the analyte result}

\item{username}{The username to log in PFS}

\item{password}{The password to log in PFS}
}
\value{
Return one data frame including project barcode and corresponding pages
}
\description{
Get Sample Test Date from PFS
}
\examples{
GET_SAMP_TestDate("Test", "user", "password");
}
