% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{Get_theta_0}
\alias{Get_theta_0}
\title{Estimate the value of theta_0 using MLE with ridge penalty}
\usage{
Get_theta_0(
  nodesLabel,
  nodesNeighbor,
  w,
  W,
  lambda = 0.1,
  threshold = 1e-04,
  maxite = 100
)
}
\arguments{
\item{nodesLabel}{The vector of latent state of all genes.}

\item{nodesNeighbor}{Neighbors of corresponding nodes in the network.}

\item{w}{The parameter in the prior for network.}

\item{W}{Another way to input the weight. For internal use.}

\item{lambda}{The tuning parameter for ridge.}

\item{threshold}{The threshold for convergence.}

\item{maxite}{The maximum number of iteration.}
}
\value{
The estimated value for theta_0:\code{beta0_new}
}
\description{
Estimate the value of theta_0 using MLE with ridge penalty
}
