% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{Get_theta_1_Gamma}
\alias{Get_theta_1_Gamma}
\title{Estimate for estimating data parameters theta_1}
\usage{
Get_theta_1_Gamma(nodesLabel, Y, N, mu, beta_1_init)
}
\arguments{
\item{nodesLabel}{Latent state of all genes.}

\item{Y}{The vector of DNM counts for all genes.}

\item{N}{The parameter in the prior for network.}

\item{mu}{The vector of mutability for all genes.}

\item{beta_1_init}{The initiate value for Newton's method.}
}
\value{
The estimated value for theta_1:\code{beta_1_new}
}
\description{
Estimate for estimating data parameters theta_1
}
