% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{Gibbs.post}
\alias{Gibbs.post}
\title{Calculate the Gibbs posterior}
\usage{
Gibbs.post(
  theta_0,
  theta_1,
  nodesNeighbor,
  w,
  W = NULL,
  Y,
  N,
  mu,
  mcmc_samples = 1010,
  burnin = 10,
  thin = 1,
  initLabel = NULL
)
}
\arguments{
\item{theta_0}{The estimated value of theta_0.}

\item{theta_1}{The estimated value of theta_1.}

\item{nodesNeighbor}{Neighbors of corresponding nodes in the network.}

\item{w}{The parameter in the prior for network.}

\item{W}{Another way to input the weight. For internal use.}

\item{Y}{The vector of DNM counts for all genes.}

\item{N}{The parameter in the prior for network.}

\item{mu}{The vector of mutability for all genes.}

\item{mcmc_samples}{The number of iterations for mcmc.}

\item{burnin}{The number of burn-ins for mcmc.}

\item{thin}{The number of thinning for mcmc.}

\item{initLabel}{The vector of initial labels for all genes.}
}
\value{
The list containing \code{S_mcmc} the mcmc samples after burning-in and thinning,
\code{q_0} the posterior probability for calculating FDR,
\code{llk} the likelihood for checking convergence
}
\description{
Calculate the Gibbs posterior
}
