% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{Gibbs.trueParam}
\alias{Gibbs.trueParam}
\title{Functions for sampling S_i from prior distribution}
\usage{
Gibbs.trueParam(
  true_theta_0,
  nodesNeighbor,
  w,
  W = NULL,
  mcmc_samples = 1000,
  burnin = 500,
  S_init = NULL
)
}
\arguments{
\item{true_theta_0}{The true value of theta_0.}

\item{nodesNeighbor}{Neighbors of corresponding nodes in the network.}

\item{w}{The parameter in the prior for network.}

\item{W}{Another way to input the weight. For internal use.}

\item{mcmc_samples}{The number of iterations for mcmc.}

\item{burnin}{The number of burn-ins for mcmc.}

\item{S_init}{The vector of initial labels for all genes.}
}
\value{
The list containing \code{S} all mcmc samples,
\code{S_mcmc} the mcmc samples after burning-in,
\code{llk} the likelihood for checking convergence
}
\description{
Functions for sampling S_i from prior distribution
}
