% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{icm_randsearch}
\alias{icm_randsearch}
\title{Estimate the parameters iteratively}
\usage{
icm_randsearch(
  nodesNeighbor,
  w = NULL,
  W = NULL,
  Y,
  mu,
  N,
  theta_0_init,
  theta_1_init,
  maxiter = 100,
  threshold = 0.001,
  lambda = 0.5,
  initLabel = NULL
)
}
\arguments{
\item{nodesNeighbor}{Neighbors of corresponding nodes in the network.}

\item{w}{The parameter in the prior for network.}

\item{W}{Another way to input the weight. For internal use.}

\item{Y}{The vector of DNM counts for all genes.}

\item{mu}{The vector of mutability for all genes.}

\item{N}{The parameter in the prior for network.}

\item{theta_0_init}{The initial value of theta_0.}

\item{theta_1_init}{The initial value of theta_1.}

\item{maxiter}{The maximum number of iteration.}

\item{threshold}{The threshold for convergence.}

\item{lambda}{The tuning parameter for estimation of theta_0.}

\item{initLabel}{The vector of initial labels for all genes.}
}
\value{
The list containing the estimated values for \code{theta_0_est},\code{theta_1_est}
and the pseudo-likelihood \code{pseudollk}.
}
\description{
Estimate the parameters iteratively
}
