% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{likpw}
\alias{likpw}
\title{Calculate the density for each node given its label}
\usage{
likpw(nodesLabel, Y, mu, N, theta_1 = NULL)
}
\arguments{
\item{nodesLabel}{The vector of latent state of all genes.}

\item{Y}{The vector of DNM counts for all genes.}

\item{mu}{The vector of mutability for all genes.}

\item{N}{The parameter in the prior for network.}

\item{theta_1}{The value of theta_1.}
}
\value{
The vector of density for all genes.
}
\description{
Calculate the density for each node given its label
}
