% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOGIN.R
\name{login_PFS}
\alias{login_PFS}
\title{Login}
\usage{
login_PFS(site, username, password)
}
\arguments{
\item{site}{The PFS version need to work on: "Test" or "Production"}

\item{username}{The username to log in PFS}

\item{password}{The password to log in PFS}
}
\value{
Return connection result
}
\description{
Login Info check and Record
}
\examples{
login_PFS(site, username, password);
}
