% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{MRF.condProb}
\alias{MRF.condProb}
\title{Calculate the conditional probability: P(Si|S_Ni,theta_0)}
\usage{
MRF.condProb(nodei, nodesLabel, nodesNeighbor, w, W, theta_0)
}
\arguments{
\item{nodei}{An indicator of node.}

\item{nodesLabel}{The vector of latent state of all genes.}

\item{nodesNeighbor}{Neighbors of corresponding nodes in the network.}

\item{w}{The parameter in the prior for network.}

\item{W}{Another way to input the weight. For internal use.}

\item{theta_0}{The value of theta_0.}
}
\value{
The conditional probability: P(Si|S_Ni,theta_0)
}
\description{
Calculate the conditional probability: P(Si|S_Ni,theta_0)
}
