% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OS_to_LIMS.R
\name{OS_to_LIMS}
\alias{OS_to_LIMS}
\title{OS Result Upload}
\usage{
OS_to_LIMS(os, site, ept_barcode, username, password)
}
\arguments{
\item{os}{The data frame that is exported from SCIEX OS analyte and read in R by READ_OS_File function}

\item{site}{The PFS version need to work on: "Test" or "Production"}

\item{ept_barcode}{The Experiment Barcode of samples}

\item{username}{The username to log in PFS}

\item{password}{The password to log in PFS}
}
\value{
Return two data frames: one is alert data frame if there is any missing or unmatched samples and the other one is the result data frame with analyte result filled in the LIMS template
}
\description{
Convert OS file to LIMS LCMS template format
}
\examples{
OS_to_LIMS("OS.txt", "Test", "LCMS1", "user", "password");
}
