% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{pseudo.llk}
\alias{pseudo.llk}
\title{Calculate the pseudo-likelihood}
\usage{
pseudo.llk(nodesLabel, nodesNeighbor, w, W, theta_0, theta_1, Y, N, mu)
}
\arguments{
\item{nodesLabel}{The vector of latent state of all genes.}

\item{nodesNeighbor}{Neighbors of corresponding nodes in the network.}

\item{w}{The parameter in the prior for network.}

\item{W}{Another way to input the weight. For internal use.}

\item{theta_0}{The value of theta_0.}

\item{theta_1}{The value of theta_1.}

\item{Y}{The vector of DNM counts for all genes.}

\item{N}{The parameter in the prior for network.}

\item{mu}{The vector of mutability for all genes.}
}
\value{
The pseudo-likelihood.
}
\description{
Calculate the pseudo-likelihood
}
