% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/READ_DTA_File.R
\name{READ_DTA_File}
\alias{READ_DTA_File}
\title{DTA Read in R}
\usage{
READ_DTA_File(data_file)
}
\arguments{
\item{data_file}{The data file after data transfer and download as .xlsx}
}
\value{
Return one data frame transferred from DTA file
}
\description{
Convert DTA file to R data frame
}
\examples{
READ_DTA_File("NP1_DTA.xlsx");
}
