% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/READ_OS_File.R
\name{READ_OS_File}
\alias{READ_OS_File}
\title{OS Result Read in R}
\usage{
READ_OS_File(OS_file)
}
\arguments{
\item{OS_file}{The data file that is exported from SCIEX OS analyte and saved as .txt}
}
\value{
Return one data frame transferred from SCIEX OS result file
}
\description{
Convert OS file to R data frame
}
\examples{
READ_OS_File("LCMS1OS.txt");
}
