% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAVE_PROJ_LotPage.R
\name{SAVE_PROJ_LotPage}
\alias{SAVE_PROJ_LotPage}
\title{Sample Lot Page Query}
\usage{
SAVE_PROJ_LotPage(site, username, password)
}
\arguments{
\item{site}{The PFS version need to work on: "Test" or "Production"}

\item{username}{The username to log in PFS}

\item{password}{The password to log in PFS}
}
\value{
Return one data frame including project barcode and corresponding pages
}
\description{
Get Sample Lot Matched Page from PFS
}
\examples{
SAVE_project_sample_lot_page("Test", "user", "password");
}
