% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCIEX_Calculation_Check.R
\name{SCIEX_calculation_check}
\alias{SCIEX_calculation_check}
\title{OS file Conversion}
\usage{
SCIEX_calculation_check(Standard, Test)
}
\arguments{
\item{Standard}{The formula file that is exported from SCIEX OS template and saved as .xlsx}

\item{Test}{The formula file that is exported from SCIEX OS analyte and saved as .xlsx}
}
\value{
Printed the result of calculation checking: whether there is a difference between the formula in the analyte file and standard file
}
\description{
Convert OS file to LIMS template format
}
\examples{
SCIEX_calculation_check(Standard, Test);
}
