% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UPDATE_ENTITY_AttrValue.R
\name{UPDATE_ENTITY_AttrValue}
\alias{UPDATE_ENTITY_AttrValue}
\title{Entity Atrribute Update}
\usage{
UPDATE_ENTITY_AttrValue(
  site,
  username,
  password,
  entity,
  barcode,
  attributes,
  values
)
}
\arguments{
\item{site}{The PFS version need to work on: "Test" or "Production"}

\item{username}{The username to log in PFS}

\item{password}{The password to log in PFS}

\item{entity}{The Entity name}

\item{barcode}{The barcode of the entity}

\item{attributes}{The vector of attribute names need to be edited}

\item{values}{The vector of updated values, order matches to attributes}
}
\value{
Return the updated information from PFS
}
\description{
Edit Entity Attributes' Value
}
\examples{
UPDATE_ENTITY_AttrValue("Test", "user", "password", "Entity", "entity123", c("A", "B"), c("1", "2"));
}
