% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UPDATE_LCMS_AttrValue.R
\name{UPDATE_LCMS_AttrValue}
\alias{UPDATE_LCMS_AttrValue}
\title{Sample Barcode Query}
\usage{
UPDATE_LCMS_AttrValue(site, username, password, ept_barcode, OS_file)
}
\arguments{
\item{site}{The PFS version need to work on: "Test" or "Production"}

\item{username}{The username to log in PFS}

\item{password}{The password to log in PFS}

\item{ept_barcode}{The Experiment Barcode of samples}

\item{OS_file}{The data file that is exported from SCIEX OS analyte and saved as .txt}
}
\value{
Return the updated information from PFS
}
\description{
Get LCMS Sample Barcode from LIMS
}
\examples{
UPDATE_LCMS_AttrValue("Test", "user", "password", "LCMS1", "LCMS1OS.txt");
}
