% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UPLOAD_OS.R
\name{UPLOAD_OS}
\alias{UPLOAD_OS}
\title{OS Result Upload}
\usage{
UPLOAD_OS(os, site, username, password)
}
\arguments{
\item{os}{The data OS data frame with barcode generated by OS_to_LIMS()}

\item{site}{The PFS version need to work on: "Test" or "Production"}

\item{username}{The username to log in PFS}

\item{password}{The password to log in PFS}
}
\value{
Return upload status message
}
\description{
Upload the OS result with sample barcode to the LIMS
}
\examples{
UPLOAD_OS("Test", "LCMS1", "user", "password");
}
