% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHECK_SAMP_Error.R
\name{CHECK_SAMP_Error}
\alias{CHECK_SAMP_Error}
\title{DTA Error Check}
\usage{
CHECK_SAMP_Error(
  format,
  data,
  exp_cols,
  lloq_opt,
  lloq = NA,
  uloq = NA,
  ulunit = NA
)
}
\arguments{
\item{format}{The format of DTA data}

\item{data}{The data frame after data transfer}

\item{exp_cols}{The columns match to sample type, analyte, unit, LLOQ/ULOQ}

\item{lloq_opt}{Need to input LLOW/ULOQ Value}

\item{lloq}{LLOQ Value}

\item{uloq}{ULOQ Value}

\item{ulunit}{LLOQ/ULOQ Unit}
}
\value{
Return data frames with expected analytes and BQL values
}
\description{
Check DTA file expected analytes and BQL values
}
\examples{
CHECK_SAMP_Error(data);
}
