% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTA.R
\name{DTA}
\alias{DTA}
\title{Result to DTA}
\usage{
DTA(
  study_id,
  version = "basic",
  site,
  project,
  username,
  password,
  column_choice = c("Study ID", "Sponsor Sample Barcode", "Subject ID", "Visit Name",
    "Sample Type", "Sample Test Date", "Test Name", "Test Result", "Test Result Unit",
    "Comments")
)
}
\arguments{
\item{study_id}{The study id for current study, acquired from LIMS project}

\item{version}{The expected version of DTA format: "basic", "plus", or "user-defined", "basic" by default}

\item{site}{The PFS version need to work on: "Test" or "Production"}

\item{project}{The Project Barcode to upload the analyte result}

\item{username}{The username to log in PFS}

\item{password}{The password to log in PFS}

\item{column_choice}{The required columns needs to be exported in the report for cilent}
}
\value{
Return result data frame in the required data transfer formatting
}
\description{
Convert Sample Lot Result to DTA format
}
\examples{
DTA(study_id = "STD1234", version = "basic", site = "Test", project = "NP123", username = "user", password = "1234")
DTA(study_id = "STD1234", version = "plus", site = "Test", project = "NP123", username = "user", password = "1234");
}
