% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OS_to_LIMS.R
\name{OS_to_LIMS}
\alias{OS_to_LIMS}
\title{OS file Conversion}
\usage{
OS_to_LIMS(path, OS_file, Template_file)
}
\arguments{
\item{path}{The working directory that stores all the input data and output data}

\item{OS_file}{The data file that is exported from SCIEX OS}

\item{Template_file}{The data template needs to be filled to upload the results to the LIMS}
}
\value{
Return nothing but write a .txt file that is ready to be uploaded to the LIMS
}
\description{
Convert OS file to LIMS template format
}
\examples{
OS_to_LIMS("~/Data", "OS.txt", "LIMS.txt");
}
