% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OS_to_LIMS.R
\name{OS_to_LIMS}
\alias{OS_to_LIMS}
\title{OS file Conversion}
\usage{
OS_to_LIMS(OS_file, Template_file)
}
\arguments{
\item{OS_file}{The data file that is exported from SCIEX OS analyte and saved as .txt}

\item{Template_file}{The data template downloaded from LIMS as .txt}
}
\value{
Return two data frames: one is alert data frame if there is any missing or unmatched samples and the other one is the result data frame with analyte result filled in the LIMS template
}
\description{
Convert OS file to LIMS template format
}
\examples{
OS_to_LIMS("~/Data", "OS.txt", "LIMS.txt");
}
