% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCIEX_Calculation_Check.R
\name{SCIEX_calculation_check}
\alias{SCIEX_calculation_check}
\title{OS Formula Check}
\usage{
SCIEX_calculation_check(standard_file, test_file)
}
\arguments{
\item{standard_file}{The formula file that is exported from SCIEX OS template and saved as .txt.}

\item{test_file}{The formula file that is exported from SCIEX OS analyte and saved as .txt.}
}
\value{
Return the result of calculation checking: whether there is a difference between the formula in the analyte file and standard file. And if there are difference, return the data frame containing the different formulas from both standard file and test file.
}
\description{
Check OS Calculation Formula with Standard
}
\examples{
SCIEX_calculation_check(Standard, Test);
}
