% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCIEX_Calculation_Check.R
\name{SCIEX_calculation_check}
\alias{SCIEX_calculation_check}
\title{OS file Conversion}
\usage{
SCIEX_calculation_check(
  standard_path = "P:/Quality/Tools/SCRT13_SCIEX OS Calculation Check/Standards/",
  test_path = "P:/Quality/Tools/SCRT13_SCIEX OS Calculation Check/Test/",
  output_path = "P:/Quality/Tools/SCRT13_SCIEX OS Calculation Check/Results/",
  Standard,
  Test
)
}
\arguments{
\item{standard_path}{The directory stored the standard SCIEX OS formula .xlsx file. There is a default settig}

\item{test_path}{The directory stored the analyte SCIEX OS formula .xlsx file. There is a default seting}

\item{output_path}{The directory to store the comparison result. There is a default setting.}

\item{Standard}{The formula file that is exported from SCIEX OS template and saved as .xlsx}

\item{Test}{The formula file that is exported from SCIEX OS analyte and saved as .xlsx}
}
\value{
Printed the result of calculation checking: whether there is a difference between the formula in the analyte file and standard file
}
\description{
Convert OS file to LIMS template format
}
\examples{
SCIEX_calculation_check(Standard, Test);
}
